Change pkg/command/cf path to internal/command path
We don't expect any of these commands to be used externally to this cf
CLI plugin, so there's no point in exporting them. Moving them behind
internal makes them only available to our main package.
