+++ internal/command/formatter.go
@@ -1,4 +1,4 @@
-package cf
+package command
 
 import (
 	"bytes"
