+++ internal/command/query.go
@@ -1,4 +1,4 @@
-package cf
+package command
 
 import (
 	"context"
