+++ internal/command/tail.go
@@ -1,4 +1,4 @@
-package cf
+package command
 
 import (
 	"context"
@@ -24,20 +24,6 @@ import (
 const (
 	timeFormat = "2006-01-02T15:04:05.00-0700"
 )
-
-// Command is the interface to implement plugin commands
-type Command func(ctx context.Context, cli plugin.CliConnection, args []string, c HTTPClient, log Logger, w io.Writer)
-
-// Logger is used for outputting log-cache results and errors
-type Logger interface {
-	Fatalf(format string, args ...interface{})
-	Printf(format string, args ...interface{})
-}
-
-// HTTPClient is the client used for HTTP requests
-type HTTPClient interface {
-	Do(req *http.Request) (*http.Response, error)
-}
 
 type TailOption func(*tailOptions)
